refactor: move fastapi server to dedicated module
This was getting unweildly to maintain in one file.
