+++ src/griptape_nodes/app/app.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import binascii
 import json
 import logging
 import os
@@ -10,13 +9,9 @@ import sys
 import threading
 from pathlib import Path
 from queue import Queue
-from typing import Any, cast
+from typing import Any
 from urllib.parse import urljoin
 
-import uvicorn
-from fastapi import FastAPI, HTTPException, Request
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.staticfiles import StaticFiles
 from griptape.events import (
     EventBus,
     EventListener,
@@ -39,10 +34,11 @@ from griptape_nodes.retained_mode.events.base_events import (
     ExecutionGriptapeNodeEvent,
     GriptapeNodeEvent,
     ProgressEvent,
-    deserialize_event,
 )
 from griptape_nodes.retained_mode.events.logger_events import LogHandlerEvent
 from griptape_nodes.retained_mode.griptape_nodes import GriptapeNodes
+
+from .api import _process_api_event, start_api
 
 # This is a global event queue that will be used to pass events between threads
 event_queue = Queue()
@@ -57,14 +53,6 @@ ws_ready_event = threading.Event()
 
 # Whether to enable the static server
 STATIC_SERVER_ENABLED = os.getenv("STATIC_SERVER_ENABLED", "true").lower() == "true"
-# Host of the static server
-STATIC_SERVER_HOST = os.getenv("STATIC_SERVER_HOST", "localhost")
-# Port of the static server
-STATIC_SERVER_PORT = int(os.getenv("STATIC_SERVER_PORT", "8124"))
-# URL path for the static server
-STATIC_SERVER_URL = os.getenv("STATIC_SERVER_URL", "/static")
-# Log level for the static server
-STATIC_SERVER_LOG_LEVEL = os.getenv("STATIC_SERVER_LOG_LEVEL", "info").lower()
 
 
 class EventLogHandler(logging.Handler):
@@ -109,7 +97,7 @@ def start_app() -> None:
 
     if STATIC_SERVER_ENABLED:
         static_dir = _build_static_dir()
-        threading.Thread(target=_serve_static_server, args=(static_dir,), daemon=True).start()
+        threading.Thread(target=start_api, args=(static_dir, event_queue), daemon=True).start()
 
     _process_event_queue()
 
@@ -137,86 +125,7 @@ def _ensure_api_key() -> str:
 def _build_static_dir() -> Path:
     """Build the static directory path based on the workspace configuration."""
     config_manager = GriptapeNodes.ConfigManager()
-    return config_manager.workspace_path / config_manager.merged_config["static_files_directory"]
-
-
-def _serve_static_server(static_dir: Path) -> None:
-    """Run FastAPI with Uvicorn in order to serve static files produced by nodes."""
-    app = FastAPI()
-
-    if not static_dir.exists():
-        static_dir.mkdir(parents=True, exist_ok=True)
-
-    app.add_middleware(
-        CORSMiddleware,
-        allow_origins=[
-            os.getenv("GRIPTAPE_NODES_UI_BASE_URL", "https://app.nodes.griptape.ai"),
-            "https://app.nodes-staging.griptape.ai",
-            "http://localhost:5173",
-        ],
-        allow_credentials=True,
-        allow_methods=["OPTIONS", "GET", "POST", "PUT"],
-        allow_headers=["*"],
-    )
-
-    app.mount(
-        STATIC_SERVER_URL,
-        StaticFiles(directory=static_dir),
-        name="static",
-    )
-
-    @app.post("/static-upload-urls")
-    async def create_static_file_upload_url(request: Request) -> dict:
-        """Create a URL for uploading a static file.
-
-        Similar to a presigned URL, but for uploading files to the static server.
-        """
-        base_url = request.base_url
-        body = await request.json()
-        file_name = body["file_name"]
-        url = urljoin(str(base_url), f"/static-uploads/{file_name}")
-
-        return {"url": url}
-
-    @app.put("/static-uploads/{file_path:path}")
-    async def create_static_file(request: Request, file_path: str) -> dict:
-        """Upload a static file to the static server."""
-        if not STATIC_SERVER_ENABLED:
-            msg = "Static server is not enabled. Please set STATIC_SERVER_ENABLED to True."
-            raise ValueError(msg)
-
-        file_full_path = Path(static_dir / file_path)
-
-        # Create parent directories if they don't exist
-        file_full_path.parent.mkdir(parents=True, exist_ok=True)
-
-        data = await request.body()
-        try:
-            file_full_path.write_bytes(data)
-        except binascii.Error as e:
-            msg = f"Invalid base64 encoding for file {file_path}."
-            logger.error(msg)
-            raise HTTPException(status_code=400, detail=msg) from e
-        except (OSError, PermissionError) as e:
-            msg = f"Failed to write file {file_path} to {static_dir}: {e}"
-            logger.error(msg)
-            raise HTTPException(status_code=500, detail=msg) from e
-
-        static_url = f"http://{STATIC_SERVER_HOST}:{STATIC_SERVER_PORT}{STATIC_SERVER_URL}/{file_path}"
-        return {"url": static_url}
-
-    @app.post("/engines/request")
-    async def create_event(request: Request) -> None:
-        body = await request.json()
-        __process_api_event(body)
-
-    logging.getLogger("uvicorn").addHandler(
-        RichHandler(show_time=True, show_path=False, markup=True, rich_tracebacks=True)
-    )
-
-    uvicorn.run(
-        app, host=STATIC_SERVER_HOST, port=STATIC_SERVER_PORT, log_level=STATIC_SERVER_LOG_LEVEL, log_config=None
-    )
+    return Path(config_manager.workspace_path) / config_manager.merged_config["static_files_directory"]
 
 
 def _init_event_listeners() -> None:
@@ -271,7 +180,7 @@ async def _alisten_for_api_requests(api_key: str) -> None:
                 try:
                     data = json.loads(message)
 
-                    __process_api_event(data)
+                    _process_api_event(data, event_queue)
                 except Exception:
                     logger.exception("Error processing event, skipping.")
         except ConnectionClosed:
@@ -465,33 +374,3 @@ def __schedule_async_task(coro: Any) -> None:
         asyncio.run_coroutine_threadsafe(coro, event_loop)
     else:
         logger.warning("Event loop not available for scheduling async task")
-
-
-def __process_api_event(event: dict) -> None:
-    """Process API events and send them to the event queue."""
-    payload = event.get("payload", {})
-
-    try:
-        payload["request"]
-    except KeyError:
-        msg = "Error: 'request' was expected but not found."
-        raise RuntimeError(msg) from None
-
-    try:
-        event_type = payload["event_type"]
-        if event_type != "EventRequest":
-            msg = "Error: 'event_type' was found on request, but did not match 'EventRequest' as expected."
-            raise RuntimeError(msg) from None
-    except KeyError:
-        msg = "Error: 'event_type' not found in request."
-        raise RuntimeError(msg) from None
-
-    # Now attempt to convert it into an EventRequest.
-    try:
-        request_event: EventRequest = cast("EventRequest", deserialize_event(json_data=payload))
-    except Exception as e:
-        msg = f"Unable to convert request JSON into a valid EventRequest object. Error Message: '{e}'"
-        raise RuntimeError(msg) from None
-
-    # Add the event to the queue
-    event_queue.put(request_event)
