+++ src/griptape_nodes/bootstrap/workflow_runners/subprocess_workflow_runner.py
@@ -1,10 +1,12 @@
 import logging
 import threading
 from multiprocessing import Process, Queue
+from multiprocessing import Queue as ProcessQueue
 from pathlib import Path
 from typing import Any
 
-from griptape_nodes.app.app import _build_static_dir, _serve_static_server
+from griptape_nodes.app.api import start_api
+from griptape_nodes.app.app import _build_static_dir
 from griptape_nodes.bootstrap.workflow_runners.local_workflow_runner import LocalWorkflowRunner
 from griptape_nodes.bootstrap.workflow_runners.workflow_runner import WorkflowRunner
 
@@ -31,7 +33,8 @@ class SubprocessWorkflowRunner(WorkflowRunner):
 
         try:
             static_dir = _build_static_dir()
-            threading.Thread(target=_serve_static_server, args=(static_dir,), daemon=True).start()
+            event_queue = ProcessQueue()
+            threading.Thread(target=start_api, args=(static_dir, event_queue), daemon=True).start()
             workflow_runner = LocalWorkflowRunner(libraries)
             workflow_runner.run(workflow_path, workflow_name, flow_input, "local")
         except Exception as e:
